requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bidict==0.22.0
boto3==1.21.37
botocore==1.24.37
CacheControl==0.12.10
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
clang==5.0
click==8.1.2
cloudpickle==1.6.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==3.4.8
cycler==0.11.0
dask==2021.11.2
datasketch==1.5.7
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator==5.1.1
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
fbmessenger==6.0.0
filelock==3.6.0
fire==0.4.0
flatbuffers==1.12
fsspec==2022.3.0
future==0.18.2
gast==0.4.0
google-auth==2.6.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.45.0
h5py==3.1.0
httptools==0.4.0
huggingface-hub==0.5.1
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.3
interchange==2021.0.4
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646983213839/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1643818144432/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
jieba==0.42.1
jmespath==1.0.0
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1646994442130/work
kafka-python==2.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
locket==0.2.1
Markdown==3.3.6
matplotlib==3.3.4
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mattermostwrapper==2.2
monotonic==1.6
msgpack==1.0.3
multidict==5.2.0
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.3.5
pansi==2020.7.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd==1.2.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.0
prompt-toolkit==2.0.10
protobuf==3.20.0
psycopg2-binary==2.9.3
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py2neo==2021.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT==2.1.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.8
pyrsistent==0.18.1
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2021.3
PyYAML==6.0
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638434924971/work
questionary==1.10.0
randomname==0.1.5
rasa==3.0.4
rasa-sdk==3.1.0
redis==3.5.3
regex==2021.8.28
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rocketchat-API==1.16.0
rsa==4.8
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
sacremoses==0.0.49
sanic==21.9.3
Sanic-Cors==1.0.1
sanic-jwt==1.7.0
sanic-plugin-toolkit==1.2.1
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.4.1
sentry-sdk==1.3.1
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
SQLAlchemy==1.4.35
tabulate==0.8.9
tarsafe==0.0.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.14.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.6.1
tensorflow-hub==0.12.0
tensorflow-probability==0.13.0
tensorflow-text==2.6.0
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tokenizers==0.12.0
toolz==0.11.2
torch==1.8.1
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.64.0
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
transformers==4.18.0
twilio==6.50.1
typeguard==2.13.3
typing-extensions==3.7.4.3
typing-utils==0.1.0
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.9
uvloop==0.16.0
wcwidth==0.2.5
webexteamssdk==1.6
websockets==10.2
Werkzeug==2.1.1
wrapt==1.12.1
yarl==1.7.2
zipp==3.8.0
